Don't read pixels if row_bytes is 0.
In these cases, there is no image which corresponds to the PixMap or BitMap
being read.  This has been observed for `icon_bitmap` in cicn resources, since
the primary purpose of the cicn is its `icon_pixmap`.  We will crash if anyone
tries to use such a PixMap or BitMap, but that's unlikely to happen.